fixed bug where clicking in swipe areas was ignored
Added a short movement detection for back_swipe and top_swipe because otherwise clicks in these areas where ignored. Since I (Quasi Kili) couldn't type the Q on the new mpos_keyboard on an esp32 touchscreen this was an urgent issue. You can test the before and after by clicking on the leftmost part of the q key in the wifi app or the upper bar in the connect 4 app.
Sending pressed AND clicked AND released events to the object below might not be the cleanest solution, but the simplest and most universal.